requirements.txt
python-telegram-bot
cloudscraper